# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Gateway template exposing user, auth, and event HTTP APIs",
        "title": "Go Gateway Template API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/v1/events/send": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Events"
                ],
                "summary": "Track an analytics event",
                "parameters": [
                    {
                        "description": "Event payload from the client",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_go-web-services_go-gateway-template_internal_dto.EventSendRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_go-web-services_go-gateway-template_internal_dto.EventSendOutputDTO"
                        }
                    }
                }
            }
        },
...
        "/v1/users/update": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Update user information.",
                "parameters": [
                    {
                        "description": "Update user request",
                        "name": "SendRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_go-web-services_go-gateway-template_internal_dto.UpdateUserInputDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_go-web-services_go-gateway-template_internal_dto.UpdateUserOutputDTO"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "enum.UserStatus": {
            "type": "string",
            "enum": [
                "active",
                "inactive",
                "pending",
                "frozen"
            ],
            "x-enum-varnames": [
                "StatusActive",
                "StatusInactive",
                "StatusPending",
                "StatusFrozen"
            ]
        },
...
        "github_com_go-web-services_go-gateway-template_internal_dto.EventSendOutputDTO": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "github_com_go-web-services_go-gateway-template_internal_dto.EventSendRequestDTO": {
            "type": "object",
            "required": [
                "distinct_id",
                "name",
                "occurred_at",
                "payload"
            ],
            "properties": {
                "distinct_id": {
                    "type": "string",
                    "example": "anon_6ba7b810-9dad-11d1-80b4-00c04fd430c8"
                },
                "message_id": {
                    "description": "MessageID is optional; when set it must be a UUID (idempotency key). Omit entirely or use a real UUID — not Swagger’s placeholder \"string\".",
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "page_viewed"
                },
                "occurred_at": {
                    "type": "string",
                    "example": "2026-04-12T10:30:00.123Z"
                },
                "payload": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "session_id": {
                    "type": "string",
                    "example": "sess_01j8xyz"
                }
            }
        },
...
        "github_com_go-web-services_go-gateway-template_internal_dto.UpdateUserInputDTO": {
            "type": "object",
            "properties": {
                "new_password": {
                    "type": "string",
                    "maxLength": 255
                },
                "old_password": {
                    "type": "string",
                    "maxLength": 255
                },
                "username": {
                    "type": "string",
                    "maxLength": 64
                }
            }
        },
        "github_com_go-web-services_go-gateway-template_internal_dto.UpdateUserOutputDTO": {
            "type": "object",
            "properties": {
                "user": {
                    "$ref": "#/definitions/github_com_go-web-services_go-gateway-template_internal_dto.UserDTO"
                }
            }
        },
        "github_com_go-web-services_go-gateway-template_internal_dto.UserDTO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "has_password": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/enum.UserStatus"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}